+++ src/memory.cpp
@@ -1,6 +1,6 @@
 
 
-func s_lin_arena make_lin_arena(size_t capacity)
+func s_lin_arena make_lin_arena(u64 capacity)
 {
 	assert(capacity > 0);
 	capacity = (capacity + 7) & ~7;
@@ -10,7 +10,7 @@ func s_lin_arena make_lin_arena(size_t capacity)
 	return result;
 }
 
-func s_lin_arena make_lin_arena_from_memory(size_t capacity, void* memory)
+func s_lin_arena make_lin_arena_from_memory(u64 capacity, void* memory)
 {
 	assert(capacity > 0);
 	assert(memory);
@@ -22,17 +22,17 @@ func s_lin_arena make_lin_arena_from_memory(size_t capacity, void* memory)
 	return result;
 }
 
-func void* la_get(s_lin_arena* arena, size_t in_requested)
+func void* la_get(s_lin_arena* arena, u64 in_requested)
 {
 	assert(in_requested > 0);
-	size_t requested = (in_requested + 7) & ~7;
+	u64 requested = (in_requested + 7) & ~7;
 	assert(arena->used + requested <= arena->capacity);
 	void* result = (u8*)arena->memory + arena->used;
 	arena->used += requested;
 	return result;
 }
 
-func void* la_get_zero(s_lin_arena* arena, size_t in_requested)
+func void* la_get_zero(s_lin_arena* arena, u64 in_requested)
 {
 	void* result = la_get(arena, in_requested);
 	memset(result, 0, in_requested);
